fix dashboard
